requirements.txt
django==4.2
djangorestframework==3.14
psycopg2-binary==2.9
django-environ==0.11
djangorestframework-simplejwt==5.3
celery==5.3
redis==5.0
django-filter==23.5
drf-spectacular==0.27.1